Fix CI build break: extract SSRF guard out of server-only module
lib/social-thumbnail.ts is imported by a client component
(app/admin/content/components/SocialPostForm.tsx, for isCdnUrl/isStorageUrl).
Importing isBlockedTarget from lib/og-scrape.ts pulled that file's
`import 'server-only'` into the client bundle, breaking the Turbopack build:
"'server-only' cannot be imported from a Client Component module."

Moved isBlockedTarget/safeFetch into lib/ssrf-guard.ts (pure functions, no
'server-only'), imported by both og-scrape.ts and social-thumbnail.ts.

Verified: npm run build succeeds with CI's placeholder env vars (previously
failed with the exact error from the PR #517 Build job).
